:zap: getOptions > refactor code for reuse purpose
